+++ src/wrap_concat.ts
@@ -1,7 +1,7 @@
 import { concat } from './concat'
 import { wrap_a_ } from './wrap_a_'
 import type { wrap_a_T } from './wrap_a_T'
-import { wrap_a2_ } from './wrap_a2_'
+import { wrap_aa_ } from './wrap_aa_'
 import type { a_nowrap_T } from './a_nowrap_T'
 import type { wrap_a_T2 } from './wrap_a_T2'
 /**
@@ -14,7 +14,7 @@ export function wrap_concat<I extends unknown = unknown>(
 	return (
 		concat<I>(
 			wrap_a_<I>(a_unwrap) as I[],
-			...(wrap_a2_<I>(rest_a) as I[]|I[][])
+			...(wrap_aa_<I>(rest_a) as I[]|I[][])
 		) as wrap_a_T<I>
 	)
 }